Move to fixed sizes for the module state.
This eliminates the string pointers in favor of fixed sizes as well as
imposing a maximum number of modules/sensors. I think this is a good
starting place for now and then if we need dynamic allocation we
can introduce that. There will be other opportunities for freeing memory
in our various buffer areas.