Started to refactor WiFi handling so we can be more async.
This is necessary for a few reasons, mainly so that we can send more
data from the SD card w/o blocking the main loop. It'll also be useful
if we ever need to use DMA or something. Basically any scenario where we
need to free up the main loop.

This is probably buggy.