Send metadata at the beginning of a binary data reading.
This is kind of a hack, though one I think I could have justified from the
beginning if I'd have anticipated this. The problem is the stateless
nature of the consumer not knowing the sensors when they're parsed. This
was just one way to solve it. So we just warn them about the data, for now.