Fix 13/11 TwoWire bus.
Noticed a ton of frequent hanging when trying to talk to the BNO055 on the
Naturalist board. This was due to the pinMode call for pin 13, from what
I can tell. Basically the pin configuration changed and that broke I2C.

See this commit. Things are much more stable w/o this.

https://github.com/arduino/ArduinoCore-samd/commit/33efce53f509e276f8c7e727ab425ed7427e9bfd