Experiment in breaking protocol handling apart from the query/replies.
One of our memory overheads is having several copies of the query/reply
structs spread across several tasks. This attempts to mitigate that by
inverting that. So we have protocol classes that have methods for reparing
and parsing the query/reply structs and then we "walk" over those. One
nice thing here is that we can make them smarter. This is one step
closer to having no ActiveObjects.