Clear cached GPS fix on every now and then, require > 1 satellites.
I saw a situation where over time a valid fix turned up because TinyGPS will
never clear a value when bad NMEA's come over. So this is to avoid that.

Example of a bad fix:

Time(1538619262) Sats(0) Hdop(0) Loc(10.300000, 0.016667, 0.000000)

So the device took that time which was hours in the past and screwed things up.

See card FK-477 for more information.